Screenshot of Audacity file in progress
This screenshot shows how I added the bell sounds to my track. I ended up looping part of the song and fading that out at the end of the track, but timing the loop correctly caused gaps in audio. I filled these in by adding a bell sound I found online under CC. I layered the sound 3 times and pitched each sound to form a note in the D major chord (I did F#, A, and D in ascending order), because that is the chord in the song at this moment.